[layout] Set a generic layouting system for all current dashboard menu
